% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{hapRFN}
\alias{hapRFN}
\title{Identifying IBD segments.}
\usage{
hapRFN(fileName, prefixPath = "", sparseMatrixPostfix = "_mat.txt",
  annotationPostfix = "_annot.txt", individualsPostfix = "_individuals.txt",
  infoPostfix = "_info.txt", labelsA = NULL, pRange = "", samples = 0,
  lowerBP = 0, upperBP = 0.05, p = 10, iter = 40, quant = 0.01,
  eps = 1e-05, l1 = 0, alpha = 0.03, cyc = 50, non_negative = 1,
  write_file = 0, norm = 0, lap = 100, IBDsegmentLength = 50,
  Lt = 0.1, Zt = 0.2, thresCount = 1e-05, mintagSNVsFactor = 3/4,
  pMAF = 0.03, haplotypes = FALSE, cut = 0.8, procMinIndivids = 0.1,
  thresPrune = 0.001, simv = "minD", minTagSNVs = 6, minIndivid = 2,
  avSNVsDist = 100, SNVclusterLength = 100, useGpu = FALSE, gpuId = -1)
}
\arguments{
\item{fileName}{The base name of the input file.}

\item{prefixPath}{The path to the input file. Default is the current directory.}

\item{sparseMatrixPostfix}{The posftix for the input matrices.
Default = "_mat.txt".}

\item{annotationPostfix}{The postfix for the annotation files.
Default value = "_annot.txt".}

\item{individualsPostfix}{The postfix for the files containing
the individuals. Default value = "_individuals.txt".}

\item{infoPostfix}{The postfix for the info file.
Default value = "_info.txt".}

\item{labelsA}{Individual names as matrix individuals x 4.}

\item{samples}{Vector of samples using one-based indexing,
representing the samples that are read. Zero means read all samples.
Default = 0.}

\item{lowerBP}{Lower bound for filtering the inputs columns, minimal MAF 
(however more than one occurence to remove private SNVs).
Default = 0.0.}

\item{upperBP}{Upper bound for filtering the inputs columns, minimal MAF.
Default = 0.05.}

\item{p}{Number of biclusters per iteration. Default = 10.}

\item{iter}{Number iterations. Default = 40.}

\item{quant}{Percentage of Ls to remove in each iteration.}

\item{eps}{Lower bound for variational parameter lapla. Default = 1e-5.}

\item{l1}{l1 weight decay. Default = 0.0.}

\item{alpha}{Sparseness loadings. Default = 0.03.}

\item{cyc}{Number of iterations. Default = 50.}

\item{non_negative}{Non-negative factors and loadings if non_negative.
Default = 1 (yes).}

\item{write_file}{Results are written to files (L in sparse format).
Default = 0 (not written).}

\item{norm}{Data normalization. Default = 0 (no normalization).}

\item{lap}{Minimal value of the variational parameter. Default = 100.0.}

\item{IBDsegmentLength}{IBD segment length in kbp. Default = 50.}

\item{Lt}{Percentage of largest Ls to consider for IBD segment extraction.
Default = 0.1.}

\item{Zt}{Percentage of largest Zs to consider for IBD segment extraction.
Default = 0.2.}

\item{thresCount}{p-value of random histogram hit. Default = 1e-5.}

\item{mintagSNVsFactor}{Percentage of segments overlap in IBD segments.
Use 1/2 for large to 3/4 for small intervals. Default = 3/4.}

\item{pMAF}{Averaged and corrected minor allele frequency. Default = 0.03.}

\item{haplotypes}{TRUE if matrix contains haplotype data. Haplotypes are
phased genotypes, so two chromosomes are expected per individual.
Default = FALSE.}

\item{cut}{Default = 0.8.}

\item{procMinIndivids}{Default = 0.1.}

\item{thresPrune}{Default = 0.001.}

\item{simv}{Default = "minD".}

\item{minTagSNVs}{Default = 6.}

\item{minIndivid}{Default = 2.}

\item{avSNVsDist}{Default = 100.}

\item{SNVclusterLength}{Default = 100.}

\item{useGpu}{TRUE to use GPU. Use this parameter together with gpuId.
Default = FALSE.}

\item{gpuId}{the ID of the GPU. This is used, when useGpu is TRUE.
Default = -1 (invalid GPU ID).}

\item{intervalSize}{Size of the interval of each split.
Default value = 10000.}

\item{shiftSize}{Size of the distance between the beginning of the
intervals. Should be smaller than the interval size, otherwise
features are skipped between intervals. Default value = 5000.}

\item{prange}{For intervals indicates its range.}
}
\value{
None
}
\description{
Identifying short IBD segments using RFNs.
}
\details{
Identifying short identity by descent (IBD) segments using Rectified
  Factor Networks (RFNs).
}
\references{
D. Clevert et al.,
  \sQuote{Rectified Factor Networks},
  CoRR, 2015 , abs/1502.06464.

  D. Clevert et al.,
  \sQuote{Rectified Factor Networks for Biclustering}
  Proceedings ICLR 2016, 2016
}
\seealso{
\code{\link{RFN}}
}
