% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{identifyDuplicates}
\alias{identifyDuplicates}
\title{Identify duplicates.}
\usage{
identifyDuplicates(fileName, startRun = 1, endRun, shiftSize = 5000,
  intervalSize = 10000)
}
\arguments{
\item{fileName}{The name of the Rda result file.}

\item{startRun}{The index of the first interval.}

\item{endRun}{The index of the last interval. Required.}

\item{shiftSize}{Size of the distance between the beginning of the
intervals. Should be smaller than the interval size, otherwise
features are skipped between intervals. Default value = 5000.}

\item{intervalSize}{Size of the interval of each split.
Default value = 10000.}
}
\value{
None
}
\description{

}
