% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{plotIBDSegment}
\alias{plotIBDSegment}
\title{IBD segment plotter}
\usage{
plotIBDSegment(x, fileName, interactive = interactive(), ...)
}
\arguments{
\item{x}{an IBD segment}

\item{fileName}{the name of the sparse matrix file that contains the
IBD segment list}

\item{interactive}{whether to prompt for new page. Default = interactive()}
}
