% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{plotIBDSegmentList}
\alias{plotIBDSegmentList}
\title{IBD segment list plotter}
\usage{
plotIBDSegmentList(x, fileName, interactive = interactive(), ...)
}
\arguments{
\item{x}{an IBDsegmentList}

\item{fileName}{the name of the sparse matrix file that contains the
IBD segment list}

\item{interactive}{whether to prompt for new page. Default = interactive()}
}
