% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{samplesPerFeature}
\alias{samplesPerFeature}
\title{Samples with non-zero feature.}
\usage{
samplesPerFeature(X, samples = 0, lowerB = 0, upperB = 1000)
}
\arguments{
\item{X}{The name of the file, containing a matrix in sparse format.}

\item{samples}{Vector of samples using one-based indexing,
representing the samples that are read. Zero means read all samples.
Default = 0.}

\item{lowerB}{Lower bound for filtering the input columns, the minimal
column sum. Default = 0.0.}

\item{upperB}{Upper bound for filtering the input columns, the maximal
column sum. Default = 1000.0.}
}
\value{
A list with elements
  \item{sL}{List with one element per feature: each element is a vector
    of samples where the feature is not zero.}
  \item{nsL}{Vector of feature length containing number of samples having
    a non-zero feature value.}
}
\description{
\code{samplesPerFeature} returns the samples for each feature, where
  this is non-zero
}
\details{
Supplies the samples for which a feature is not zero.

The data matrix is directly scanned by the C code and must be in sparse
  matrix format.

Sparse matrix format where rows are as follows:
\itemize{
  \item number of rows (\eqn{nrow})
  \item number of non-zero elements (\eqn{nnz})
  \item row pointers, a list of integers of length \eqn{nrow + 1},
    where the ith element is the number of non-zero elements 
    until the \eqn{i - 1}th row
  \item column indices, a list of integers of length \eqn{nnz}, 
    where the \eqn{i}th element is the (zero based) column index 
    of the \eqn{i}th non-zero element
  \item values, the list of non-zero elements
}
}
