% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{analyzeIBDsegments}
\alias{analyzeIBDsegments}
\title{Analyze IBD segments.}
\usage{
analyzeIBDsegments(fileName, startRun = 1, endRun = 0, shiftSize = 5000,
  intervalSize = 10000)
}
\arguments{
\item{fileName}{The name of the Rda result file.}

\item{startRun}{The index of the first interval. Default = 1}

\item{endRun}{The index of the last interval. Default = 0, which means the last index.}

\item{shiftSize}{Size of the distance between the beginning of the
intervals. Should be smaller than the interval size, otherwise
features are skipped between intervals. Default value = 5000.}

\item{intervalSize}{Size of the interval of each split.
Default value = 10000.}
}
\value{
A list with elements
  \item{startRun}{The parameter startRun.}
  \item{endRun}{The parameter endRun.}
  \item{noIBDsegments}{The number of IBD segments.}
  \item{avIBDsegmentPos}{Vector of the genomic locations of the IBD segments.}
  \item{avIBDsegmentLengthSNV}{Vector of lengths in SNVs of the IBD segments.}
  \item{avIBDsegmentLength}{Vector of lengths in bp of the IBD segments.}
  \item{avnoIndivid}{Vector of number of samples belonging to the IBD segments.}
  \item{avnoTagSNVs}{Vector of number of tag SNVs marking the IBD segments.}
  \item{avnoFreq}{Vector of frequencies of tagSNVs in the whole dataset.}
  \item{avnoGroupFreq}{Vector of frequencies of tagSNVs in the population
    that is considered.}
  \item{avnotagSNVChange}{Vector of flags indicating a switch between minor
    and major alleles of tagSNVs (1 means switched, 0 not switched).}
  \item{avnotagSNVsPerIndividual}{Vector of number of tagSNVs per individual.}
  \item{avnoindividualPerTagSNV}{Vector of number of individuals that posses
    the minor allele per tagSNV.}
  \item{avIBDsegmentPosS}{Summary of avIBDsegmentPos.}
  \item{avIBDsegmentLengthSNVS}{Summary of avIBDsegmentLengthSNV.}
  \item{avIBDsegmentLengthS}{Summary of avIBDsegmentLength.}
  \item{avnoIndividS}{Summary of avnoIndivid.}
  \item{avnoTagSNVsS}{Summary of avnoTagSNVs.}
  \item{avnoFreqS}{Summary of avnoFreq.}
  \item{avnoGroupFreqS}{Summary of avnoGroupFreq.}
  \item{avnotagSNVChangeS}{Summary of avnotagSNVChange.}
  \item{avnotagSNVsPerIndividualS}{Summary of avnotagSNVsPerIndividual.}
  \item{avnoindividualPerTagSNVS}{Summary of avnoindividualPerTagSNV.}
}
\description{
\code{analyzeIBDsegments} loops over all intervals and provides
  statistics about the IBD segments.
}
\details{
The function loops over all intervals and provides statistics about the
  IBD segments. The loop goes over the intervals that have been provided by
  \code{iterateIntervals}. Duplicates are ignore at this analysis and must
  be identified in a preceding step via \code{identifyDuplicates}.
  Other statistics and annotations can be computed if the code is changed
  accordingly.

  Results are written to the file "analyzeResults.Rda".
}
\seealso{
\code{\link{iterateIntervals}} and
  \code{\link{identifyDuplicates}}
}
