% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{identifyDuplicates}
\alias{identifyDuplicates}
\title{Identify duplicates.}
\usage{
identifyDuplicates(fileName, startRun = 1, endRun = 0, shiftSize = 5000,
  intervalSize = 10000)
}
\arguments{
\item{fileName}{The name of the Rda result file.}

\item{startRun}{The index of the first interval. Default = 1}

\item{endRun}{The index of the last interval. Default = 0, which means the last index.}

\item{shiftSize}{Size of the distance between the beginning of the
intervals. Should be smaller than the interval size, otherwise
features are skipped between intervals. Default value = 5000.}

\item{intervalSize}{Size of the interval of each split.
Default value = 10000.}
}
\value{
None
}
\description{
\code{identifyDuplicates} identifies IBD segments that are similar to
  each other. This function is used in combination with \code{vcf2sparse}
  which splits a VCF file into overlapping intervals. These intervals are
  then analyzed by \code{analyzeIntervals}, and the checked for duplicates
  by this function.

  Results are written to the file "dups.Rda".
}
