% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapRFN.R
\name{readSparseSamples}
\alias{readSparseSamples}
\title{Read sparse matrix samples.}
\usage{
readSparseSamples(X, samples = 0, lowerB = 0, upperB = 1000)
}
\arguments{
\item{X}{The name of the file, containing a matrix in sparse format.}

\item{samples}{Vector of samples using one-based indexing,
representing the samples that are read. Zero means read all samples.
Default = 0.}

\item{lowerB}{Lower bound for filtering the input columns, the minimal
column sum. Default = 0.0.}

\item{upperB}{Upper bound for filtering the input columns, the maximal
column sum. Default = 1000.0.}
}
\value{
Data matrix of given samples
}
\description{
\code{readSparseSamples} Reads sparse matrix samples.
}
\details{
Reads the sparse matrix samples filtered by \code{samples}, \code{lowerB}
.  and \code{upperB}.

The data matrix is directly scanned by the C code and must be in sparse
  matrix format.

Sparse matrix format where rows are as follows:
\itemize{
  \item number of non-zero elements (\eqn{nnz})
  \item number of rows (\eqn{nrow})
  \item row pointers, a list of integers of length \eqn{nrow + 1},
    where the ith element is the number of non-zero elements 
    until the \eqn{i - 1}th row
  \item column indices, a list of integers of length \eqn{nnz}, 
    where the \eqn{i}th element is the (zero based) column index 
    of the \eqn{i}th non-zero element
  \item values, the list of non-zero elements
}
}
